/v1/notice/list: 공지사항 페이징&검색어 처리, 공지사항 hit 고치기 (db 밀고 다시)

+++ backend/src/main/java/place/skillexchange/backend/repository/NoticeRepository.java
@@ -1,5 +1,8 @@
 package place.skillexchange.backend.repository;
 
+import com.querydsl.core.types.dsl.BooleanExpression;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import place.skillexchange.backend.entity.File;
 import place.skillexchange.backend.entity.Notice;
@@ -7,8 +10,9 @@ import place.skillexchange.backend.entity.Notice;
 import java.util.List;
 import java.util.Optional;
 
-public interface NoticeRepository extends JpaRepository<Notice,Integer> {
+public interface NoticeRepository extends JpaRepository<Notice,Long>{
     Optional<Notice> findById(Long noticeId);
 
     void deleteById(Long noticeId);
 }
+
